criptografar senha do usuario

+++ src/main/java/br/com/filmesplus/avaliacao/respositories/AvaliacaoRepo.java
@@ -3,5 +3,5 @@ package br.com.filmesplus.avaliacao.respositories;
 import br.com.filmesplus.avaliacao.models.Avaliacao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AvaliacaoRepo extends JpaRepository<Avaliacao, Integer> {
+public interface AvaliacaoRepo extends JpaRepository<Avaliacao, Long> {
 }
